cloud_functions/cbs/deaths/requirements.txt
pandas==1.3.4
requests==2.26.0
google-cloud-storage==1.43.0
openpyxl==3.0.9
